# Fix possible bugs in passthrough_flat_pore, optimize code